// src/app/interfaces/movie.ts
import { Person } from "./person";

export interface Movie {
    id: string;
    cast: string[];
    director: string;
    directorId: string;
    genres: string[];
    imageUrl: string;
    releaseYear: number;
    score: number;
    summary: string;
    title: string;
    castFullDetails?: Person[];
}